% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_data.R
\name{bin_data}
\alias{bin_data}
\title{Create dataframe of binned counts}
\usage{
bin_data(z_vector, binv, zstar, binwidth, bins_l, bins_r)
}
\arguments{
\item{z_vector}{the vector of (unbinned) data to be analysed.}

\item{binv}{zstar location within its bin (min, max or median value). Default is median.}

\item{zstar}{the bunching point.}

\item{binwidth}{width of each bin.}

\item{bins_l}{bins to left of zstar to use.}

\item{bins_r}{bins to right of zstar to use.}
}
\value{
A dataframe with the bin value and counts per bin
}
\description{
Create dataframe of binned counts
}
\seealso{
\code{\link{bunchit}}
}
