% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bunching_data}
\alias{bunching_data}
\title{Simulated data for bunching examples.}
\format{
A data frame with 27510 rows and 2 variables:
\describe{
  \item{kink_vector}{simulated earnings vector, suitable for examples of bunching at kinks}.
  \item{notch_vector}{simulated earnings vector, suitable for examples of bunching at notches}.
}
}
\usage{
bunching_data
}
\description{
A dataset containing two simulated vectors of about 27,500 observations.
}
\seealso{
\code{\link{bunching}}, \code{\link{bunchit}}
}
\keyword{datasets}
