% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bunchit.R
\name{bunchit}
\alias{bunchit}
\title{Run the bunching estimator}
\usage{
bunchit(z_vector, binv = "median", zstar, binwidth, bins_l, bins_r, poly,
  bins_excl_l, bins_excl_r, extra_fe = NA, rn = NA, n_boot = 50,
  correct = T, iter_max = 200, t0, t1, T0 = 0, force_notch = F,
  p_title = "", p_xtitle = "z_name", p_ytitle = "Count",
  p_maxy = NA, p_axis_txt_size = 7, p_axis_val_size = 7,
  p_theme = "bw_light", p_freq_color = "black",
  p_cf_color = "maroon", p_zstar_color = "red", p_freq_size = 0.5,
  p_cf_size = 0.5, p_freq_msize = 1, p_zstar_size = 0.5, p_b = T,
  p_b_xpos = posx, p_b_ypos = posy, p_b_size = 3)
}
\arguments{
\item{z_vector}{the vector of (unbinned) data to be analysed.}

\item{binv}{zstar location within its bin (min, max or median value). Default is median.}

\item{zstar}{the bunching point.}

\item{binwidth}{width of each bin.}

\item{bins_l}{bins to left of zstar to use.}

\item{bins_r}{bins to right of zstar to use.}

\item{poly}{order of polynomial for counterfactual fit.}

\item{bins_excl_l}{number of bins left of zstar to include in bunching region.}

\item{bins_excl_r}{number of bins right of zstar to include in bunching region.}

\item{extra_fe}{bin values to control for using fixed effect.}

\item{rn}{round number bunching (up to 2 round numbers).}

\item{n_boot}{number of bootstrapped iterations.}

\item{correct}{whether to implement correction for integration constraint.}

\item{iter_max}{maximum iterations for integration constraint correction.}

\item{t0}{Marginal tax rate below zstar.}

\item{t1}{Marginal tax rate above zstar.}

\item{T0}{size of notch (for kink, T0 = 0).}

\item{force_notch}{whether to enforce manual choice of zu.}

\item{p_title}{plot title.}

\item{p_xtitle}{plot x_axis label.}

\item{p_ytitle}{plot y_axis label.}

\item{p_maxy}{plot's maximum y_axis value.}

\item{p_theme}{plot theme.}

\item{p_freq_color}{plot's frequency line color}

\item{p_cf_color}{plot's counterfactual line color}

\item{p_zstar_color}{plot's bunching region marker lines color}

\item{p_freq_size}{plot's frequency line thickness}

\item{p_cf_size}{plot's counterfactual line thickness}

\item{p_zstar_size}{plot's bunching region marker lines thickness}

\item{p_b}{Should bunching estimate be shown on plot?}

\item{p_b_xpos}{plot's xaxis coordinate of bunching estimate}

\item{p_b_ypos}{plot's yaxis coordinate of bunching estimate}

\item{p_b_size}{size of plot's printed bunching estimate}

\item{p_txt_size}{text size of plot' axes' labels.}

\item{p_cf_msize}{plot's counterfactual markers' size}
}
\description{
Run the bunching estimator
}
