% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticity.R
\name{elasticity}
\alias{elasticity}
\title{Estimate elasticity from single bunching observation.}
\usage{
elasticity(beta, binwidth, zstar, t0, t1)
}
\arguments{
\item{beta}{normalised excess mass.}

\item{binwidth}{width of each bin.}

\item{zstar}{the bunching point.}

\item{t0}{Marginal tax rate below zstar.}

\item{t1}{Marginal tax rate above zstar.}
}
\value{
Elasticity estimate.
}
\description{
Estimate elasticity from single bunching observation.
}
\seealso{
\code{\link{bunchit}}
}
