% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticity_sd.R
\name{elasticity_sd}
\alias{elasticity_sd}
\title{Elasticity s.d.
Estimate standard deviation of elasticity.}
\usage{
elasticity_sd(b_vector, binwidth, zstar, t0, t1, notch)
}
\arguments{
\item{b_vector}{vector of normalised excess mass estimates.}

\item{binwidth}{a numeric value for the width of each bin.}

\item{zstar}{a numeric value for the the bunching point.}

\item{t0}{numeric value between 0 and 1 setting the marginal/average tax rate below zstar, depending on kink/notch choice. see notch parameter.}

\item{t1}{numeric value between 0 and 1 setting the marginal/average tax rate above zstar, depending on kink/notch choice. see notch parameter.}

\item{notch}{if TRUE, zstar treated as notch. Default is kink.}
}
\value{
\code{elasticity_sd} returns the standard deviation of the elasticity estimates.
}
\description{
Elasticity s.d.
Estimate standard deviation of elasticity.
}
\seealso{
\code{\link{elasticity}} \code{\link{bunchit}}
}
