% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_bunching.R
\name{fit_bunching}
\alias{fit_bunching}
\title{Fit  bunching model and estimate excess mass.}
\usage{
fit_bunching(thedata, themodelformula, notch, zD_bin)
}
\arguments{
\item{thedata}{(binned) data that includes all variables necessary for fitting the model.}

\item{themodelformula}{formula to fit.}

\item{notch}{if TRUE, zstar treated as notch. Default is kink.}

\item{zD_bin}{the bin marking the upper end of the dominated region (notch case).}
}
\value{
coefficients, residuals, cf_density, bunchers_excess, cf_bunchers, b_estimate, bins_bunchers, model_formula.
}
\description{
Fit  bunching model and estimate excess mass.
}
\seealso{
\code{\link{bunchit}}, \code{\link{prep_data_for_fit}}
}
