% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notch_equation.R
\name{notch_equation}
\alias{notch_equation}
\title{Notch Equation}
\usage{
notch_equation(e, t0, t1, zstar, dzstar)
}
\arguments{
\item{e}{elasticity.}

\item{t0}{numeric value between 0 and 1 setting the marginal/average tax rate below zstar, depending on kink/notch choice. see notch parameter.}

\item{t1}{numeric value between 0 and 1 setting the marginal/average tax rate above zstar, depending on kink/notch choice. see notch parameter.}

\item{zstar}{a numeric value for the the bunching point.}

\item{dzstar}{The distance of the marginal buncher above zstar.}
}
\value{
\code{util_diff} returns the difference in utility between zstar and z_I in notch setting.
}
\description{
Defines indifference condition based on parametric utility function in notch setting.
}
\seealso{
\code{\link{bunchit}}

\code{\link{elasticity}}
}
