% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notch_indifference.R
\name{notch_indifference}
\alias{notch_indifference}
\title{Indifference Level in Notch Setting}
\usage{
notch_indifference(notch_marginal_buncher, t0, t1, e)
}
\arguments{
\item{notch_marginal_buncher}{location of marginal buncher in notch, i.e. zI for the one initially at zstar + Dzstar.}

\item{t0}{numeric value between 0 and 1 setting the marginal/average tax rate below zstar, depending on kink/notch choice. see notch parameter.}

\item{t1}{numeric value between 0 and 1 setting the marginal/average tax rate above zstar, depending on kink/notch choice. see notch parameter.}

\item{e}{The estimated elasticity (from the parametric version of the notch setting).}
}
\value{
\code{notch_indifference} returns the indifference location of the marginal buncher, i.e. zI.
}
\description{
Calculate location (value of running variable) where marginal buncher is indifferent between staying there and moving to notch.
}
\seealso{
\code{\link{bunchit}}
}
