% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bunching.R
\name{plot_bunching}
\alias{plot_bunching}
\title{Create bunching plot.}
\usage{
plot_bunching(binned_data, cf, zstar, binwidth, bins_excl_l, bins_excl_r,
  p_title, p_xtitle, p_ytitle, p_maxy, p_axis_txt_size, p_axis_val_size,
  p_theme, p_freq_color, p_cf_color, p_zstar_color, p_freq_size, p_cf_size,
  p_freq_msize, p_zstar_size, p_b, b, b_sd, p_b_xpos, p_b_ypos, p_b_size,
  t0 = NA, t1 = NA, notch = F, domregion_color = NA)
}
\arguments{
\item{binned_data}{binned data with frequency and counterfactual estimated}

\item{zstar}{the bunching point.}

\item{binwidth}{width of each bin.}

\item{bins_excl_l}{number of bins left of zstar to include in bunching region.}

\item{bins_excl_r}{number of bins right of zstar to include in bunching region.}

\item{p_title}{plot title.}

\item{p_xtitle}{plot x_axis label.}

\item{p_ytitle}{plot y_axis label.}

\item{p_maxy}{plot's maximum y_axis value.}

\item{p_theme}{plot theme.}

\item{p_freq_color}{plot's frequency line color}

\item{p_cf_color}{plot's counterfactual line color}

\item{p_zstar_color}{plot's bunching region marker lines color}

\item{p_freq_size}{plot's frequency line thickness}

\item{p_cf_size}{plot's counterfactual line thickness}

\item{p_zstar_size}{plot's bunching region marker lines thickness}

\item{p_b}{Should bunching estimate be shown on plot?}

\item{p_b_xpos}{plot's xaxis coordinate of bunching estimate}

\item{p_b_ypos}{plot's yaxis coordinate of bunching estimate}

\item{p_b_size}{size of plot's printed bunching estimate}

\item{t0}{marginal/average tax rate below zstar. see notch option.}

\item{t1}{marginal/average tax rate above zstar. see notch option.}

\item{notch}{whether it is a notch or kink. Default is kink.}
}
\value{
A plot with the frequency, counterfactual and bunching region demarcated. Can also include the bunching estimate in the plot if specified.
}
\description{
Create bunching plot.
}
\seealso{
\code{\link{bunchit}}
}
