% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hist.R
\name{plot_hist}
\alias{plot_hist}
\title{Create simple binned plot.}
\usage{
plot_hist(z_vector, binv = "median", zstar, binwidth, bins_l, bins_r,
  p_title = "", p_xtitle = "z_name", p_ytitle = "Count",
  p_maxy = NA, p_axis_txt_size = 7, p_axis_val_size = 7,
  p_theme = "bw_light", p_freq_color = "black",
  p_zstar_color = "red", p_freq_size = 0.5, p_freq_msize = 1,
  p_zstar_size = 0.5, p_zstar = T)
}
\arguments{
\item{z_vector}{a numeric vector of (unbinned) data to be analysed.}

\item{binv}{a string setting location of zstar within its bin ("min", "max" or "median" value). Default is median.}

\item{zstar}{a numeric value for the the bunching point.}

\item{binwidth}{a numeric value for the width of each bin.}

\item{bins_l}{number of bins to left of zstar to use in analysis.}

\item{bins_r}{number of bins to right of zstar to use in analysis.}

\item{p_title}{plot title.}

\item{p_xtitle}{plot x_axis label.}

\item{p_ytitle}{plot y_axis label.}

\item{p_maxy}{plot's maximum y_axis value.}

\item{p_theme}{plot theme (in ggplot2 format).}

\item{p_freq_color}{plot's frequency line color}

\item{p_zstar_color}{plot's bunching region marker lines color}

\item{p_freq_size}{plot's frequency line thickness}

\item{p_zstar_size}{plot's bunching region marker lines thickness}

\item{p_zstar}{whether to show vertical line for zstar. Default is true.}
}
\value{
A plot with just the frequency per bin (histogram).
}
\description{
Create simple binned plot.
}
\seealso{
\code{\link{bunchit}}
}
