% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data_for_reg.R
\name{prep_data_for_reg}
\alias{prep_data_for_reg}
\title{Prepare binned data and model for bunching estimation.}
\usage{
prep_data_for_reg(data_binned, zstar, binwidth, bins_l, bins_r, poly,
  bins_excl_l, bins_excl_r, rn, extra_fe)
}
\arguments{
\item{data_binned}{dataframe of counts per bin}

\item{zstar}{the bunching point.}

\item{binwidth}{width of each bin.}

\item{bins_l}{bins to left of zstar to use.}

\item{bins_r}{bins to right of zstar to use.}

\item{poly}{order of polynomial for counterfactual fit.}

\item{bins_excl_l}{number of bins left of zstar to include in bunching region.}

\item{bins_excl_r}{number of bins right of zstar to include in bunching region.}

\item{rn}{round number bunching (up to 2 round numbers).}

\item{extra_fe}{bin values to control for using fixed effect.}
}
\description{
Prepare binned data and model for bunching estimation.
}
\seealso{
\code{\link{bunchit}}
}
