% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare binned data and model for bunching estimation.}
\usage{
prepare_data(z_vector, binv, zstar, binwidth, bins_l, bins_r, poly,
  bins_excl_l, bins_excl_r, rn, extra_fe)
}
\arguments{
\item{z_vector}{the vector of (unbinned) data to be analysed.}

\item{binv}{zstar location within its bin (min, max or median value). Default is median.}

\item{zstar}{the bunching point.}

\item{binwidth}{width of each bin.}

\item{bins_l}{bins to left of zstar to use.}

\item{bins_r}{bins to right of zstar to use.}

\item{poly}{order of polynomial for counterfactual fit.}

\item{bins_excl_l}{number of bins left of zstar to include in bunching region.}

\item{bins_excl_r}{number of bins right of zstar to include in bunching region.}

\item{rn}{round number bunching (up to 2 round numbers).}

\item{extra_fe}{bin values to control for using fixed effect.}
}
\description{
Prepare binned data and model for bunching estimation.
}
\seealso{
\code{\link{bunchit}}
}
