% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_data.R
\name{bin_data}
\alias{bin_data}
\title{Create dataframe of binned counts}
\usage{
bin_data(z_vector, binv, zstar, binwidth, bins_l, bins_r)
}
\arguments{
\item{z_vector}{a numeric vector of (unbinned) data to be analysed.}

\item{binv}{a string setting location of zstar within its bin ("min", "max" or "median" value). Default is median.}

\item{zstar}{a numeric value for the the bunching point.}

\item{binwidth}{a numeric value for the width of each bin.}

\item{bins_l}{number of bins to left of zstar to use in analysis.}

\item{bins_r}{number of bins to right of zstar to use in analysis.}
}
\value{
A dataframe with the bin value and counts per bin
}
\description{
Create dataframe of binned counts
}
\seealso{
\code{\link{bunchit}}
}
