% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bunchit.R
\name{bunchit}
\alias{bunchit}
\title{Run the bunching estimator}
\usage{
bunchit(z_vector, binv = "median", zstar, binwidth, bins_l, bins_r, poly,
  bins_excl_l, bins_excl_r, extra_fe = NA, rn = NA, n_boot = 50,
  correct = T, iter_max = 200, t0, t1, notch = F, force_notch = F,
  p_title = "", p_xtitle = "z_name", p_ytitle = "Count",
  p_maxy = NA, p_axis_txt_size = 7, p_axis_val_size = 7,
  p_theme = "bw_light", p_freq_color = "black",
  p_cf_color = "maroon", p_zstar_color = "red", p_freq_size = 0.5,
  p_cf_size = 0.5, p_freq_msize = 1, p_zstar_size = 0.5, p_b = T,
  p_b_xpos = posx, p_b_ypos = posy, p_b_size = 3,
  domregion_color = "grey40", seed = NA)
}
\arguments{
\item{z_vector}{a numeric vector of (unbinned) data to be analysed.}

\item{binv}{a string setting location of zstar within its bin ("min", "max" or "median" value). Default is median.}

\item{zstar}{a numeric value for the the bunching point.}

\item{binwidth}{a numeric value for the width of each bin.}

\item{bins_l}{number of bins to left of zstar to use in analysis.}

\item{bins_r}{number of bins to right of zstar to use in analysis.}

\item{poly}{a numeric value for the order of polynomial for counterfactual fit.}

\item{bins_excl_l}{number of bins to left of zstar to include in bunching region.}

\item{bins_excl_r}{number of bins to right of zstar to include in bunching region.}

\item{extra_fe}{a numeric vector of bin values to control for using fixed effects.}

\item{rn}{a numeric vector of round numbers (up to 2) to control for.}

\item{n_boot}{number of bootstrapped iterations.}

\item{correct}{if TRUE, implements correction for integration constraint.}

\item{iter_max}{maximum iterations for integration constraint correction.}

\item{t0}{numeric value between 0 and 1 setting the marginal/average tax rate below zstar, depending on kink/notch choice. see notch parameter.}

\item{t1}{numeric value between 0 and 1 setting the marginal/average tax rate above zstar, depending on kink/notch choice. see notch parameter.}

\item{notch}{if TRUE, zstar treated as notch. Default is kink.}

\item{force_notch}{if TRUE, user choice of zu (upper limit of bunching region) is enforced. Default is FALSE (zu set by setting bunching equal to missing mass).}

\item{p_title}{plot title.}

\item{p_xtitle}{plot x_axis label.}

\item{p_ytitle}{plot y_axis label.}

\item{p_maxy}{plot's maximum y_axis value.}

\item{p_theme}{plot theme (in ggplot2 format).}

\item{p_freq_color}{plot's frequency line color}

\item{p_cf_color}{plot's counterfactual line color}

\item{p_zstar_color}{plot's bunching region marker lines color}

\item{p_freq_size}{plot's frequency line thickness}

\item{p_cf_size}{plot's counterfactual line thickness}

\item{p_zstar_size}{plot's bunching region marker lines thickness}

\item{p_b}{if TRUE, plot also includes bunching estimate. Default is TRUE.}

\item{p_b_xpos}{plot's xaxis coordinate of bunching estimate}

\item{p_b_ypos}{plot's yaxis coordinate of bunching estimate}

\item{p_b_size}{size of plot's printed bunching estimate}

\item{seed}{a numeric value for bootstrap seed (random re-sampling of residuals).}

\item{p_txt_size}{text size of plot' axes' labels.}

\item{p_cf_msize}{plot's counterfactual markers' size}

\item{p_domregion_color}{plot's dominated region marker line color (notch case).}
}
\description{
Run the bunching estimator
}
