% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bunchit.R
\name{bunchit}
\alias{bunchit}
\title{Bunching Estimator}
\usage{
bunchit(z_vector, binv = "median", zstar, binwidth, bins_l, bins_r,
  poly = 9, bins_excl_l = 0, bins_excl_r = 0, extra_fe = NA,
  rn = NA, n_boot = 100, correct = TRUE,
  correction_above_zu = FALSE, iter_max = 200, t0, t1, notch = FALSE,
  force_notch = FALSE, e_parametric = FALSE, e_parametric_lb = 1e-04,
  e_parametric_ub = 3, p_title = "", p_xtitle = "z_name",
  p_ytitle = "Count", p_title_size = 9, p_axis_title_size = 9,
  p_axis_val_size = 7.5, p_miny = 0, p_maxy = NA, p_ybreaks = NA,
  p_theme = "theme_classic()", p_freq_color = "black",
  p_cf_color = "maroon", p_zstar_color = "red",
  p_grid_major_y_color = "lightgrey", p_freq_size = 0.5,
  p_freq_msize = 1, p_cf_size = 0.5, p_zstar_size = 0.5,
  p_b = FALSE, p_e = FALSE, p_b_e_xpos = NA, p_b_e_ypos = NA,
  p_b_e_size = 3, p_domregion_color = "blue", seed = NA,
  p_domregion_ltype = "longdash")
}
\arguments{
\item{z_vector}{a numeric vector of (unbinned) data to be analysed.}

\item{binv}{a string setting location of zstar within its bin ("min", "max" or "median" value). Default is median.}

\item{zstar}{a numeric value for the the bunching point.}

\item{binwidth}{a numeric value for the width of each bin.}

\item{bins_l}{number of bins to left of zstar to use in analysis.}

\item{bins_r}{number of bins to right of zstar to use in analysis.}

\item{poly}{a numeric value for the order of polynomial for counterfactual fit.}

\item{bins_excl_l}{number of bins to left of zstar to include in bunching region.}

\item{bins_excl_r}{number of bins to right of zstar to include in bunching region.}

\item{extra_fe}{a numeric vector of bin values to control for using fixed effects.}

\item{rn}{a numeric vector of round numbers (up to 2) to control for.}

\item{n_boot}{number of bootstrapped iterations.}

\item{correct}{if TRUE, implements correction for integration constraint.}

\item{iter_max}{maximum iterations for integration constraint correction.}

\item{t0}{numeric value between 0 and 1 setting the marginal/average tax rate below zstar, depending on kink/notch choice. see notch parameter.}

\item{t1}{numeric value between 0 and 1 setting the marginal/average tax rate above zstar, depending on kink/notch choice. see notch parameter.}

\item{notch}{if TRUE, zstar treated as notch. Default is kink.}

\item{force_notch}{if TRUE, user choice of zu (upper limit of bunching region) is enforced. Default is FALSE (zu set by setting bunching equal to missing mass).}

\item{e_parametric}{if TRUE, elasticity is estimated using parametric specification (quasi-linear and iso-elastic utility function). Default is FALSE (which estimates reduced-form approximation).}

\item{e_parametric_lb}{lower bound for estimate elasticity using parametric specification for notch (quasi-linear and iso-elastic utility function).}

\item{e_parametric_ub}{upper bound for estimate elasticity using parametric specification for notch (quasi-linear and iso-elastic utility function).}

\item{p_title}{plot title.}

\item{p_xtitle}{plot x_axis label.}

\item{p_ytitle}{plot y_axis label.}

\item{p_title_size}{size of plot's title.}

\item{p_axis_title_size}{size of plot's axes' title labels.}

\item{p_axis_val_size}{size of plot's axes' numeric labels.}

\item{p_miny}{plot's minimum y_axis value. Default is zero.}

\item{p_maxy}{plot's maximum y_axis value. Default is optimized internally.}

\item{p_ybreaks}{the y-axis value(s) at which to add horizontal line markers. Default is optimized internally.}

\item{p_theme}{plot theme (in ggplot2 format).}

\item{p_freq_color}{plot's frequency line color.}

\item{p_cf_color}{plot's counterfactual line color.}

\item{p_zstar_color}{plot's bunching region marker lines color.}

\item{p_grid_major_y_color}{plot's y-axis major grid line color.}

\item{p_freq_size}{plot's frequency line thickness.}

\item{p_freq_msize}{plot's frequency line marker size.}

\item{p_cf_size}{plot's counterfactual line thickness.}

\item{p_zstar_size}{plot's bunching region marker line thickness.}

\item{p_b}{if TRUE, plot also includes bunching estimate. Default is TRUE.}

\item{p_e}{if TRUE, plot also includes elasticity estimate. Only shown if bunching estimate shown. Default is TRUE.}

\item{p_b_e_xpos}{plot's xaxis coordinate of bunching/elasticity estimate.}

\item{p_b_e_ypos}{plot's yaxis coordinate of bunching/elasticity estimate.}

\item{p_b_e_size}{size of plot's printed bunching/elasticity estimate.}

\item{p_domregion_color}{plot's dominated region marker line color (notch case).}

\item{seed}{a numeric value for bootstrap seed (random re-sampling of residuals).}

\item{p_domregion_ltype}{a string for the vertical line type marking the dominated region (zD) in the plot (notch case only).}
}
\value{
\code{bunchit} returns a list of results, both for visualizing and for further analysis of the data underlying the estimates. These include:
  \item{plot}{The bunching plot.}
  \item{data}{The binned data used for estimation.}
  \item{cf}{The estimated counterfactuals.}
  \item{B}{The estimated excess mass (not normalized).}
  \item{B_vector}{The vector of bootstrapped B's.}
  \item{B_sd}{The standard deviation of B_vector.}
  \item{b}{The estimated excess mass (normalized).}
  \item{b_vector}{The vector of bootstrapped b's.}
  \item{b_sd}{The standard deviation of b_vector.}
  \item{e}{The estimated elasticity.}
  \item{e_vector}{The vector of bootstrapped elasticities (e).}
  \item{e_sd}{The standard deviation of e_vector.}
  \item{alpha}{The estimated fraction of bunchers in dominated region (notch case).}
  \item{alpha_vector}{The vector of bootstrapped alphas.}
  \item{alpha_sd}{The standard deviation of alpha_vector.}
  \item{model_fit}{The model fit on the actual (i.e. not bootstrapped) data.}
  \item{zD}{The value demarcating the dominated region (notch case).}
  \item{zD_bin}{The bin above zstar demarcating the dominated region (notch case).}
  \item{zU_bin}{The location of zU (upper range of excluded region) as estimated from notch setting by setting force_notch = FALSE.}
  \item{marginal_buncher}{The location (z value) of the marginal buncher.}
  \item{marginal_buncher_vector}{The vector of bootstrapped marginal_buncher values.}
  \item{marginal_buncher_sd}{The standard deviation of marginal_buncher_vector.}
}
\description{
Implement the bunching estimator in a kink or notch setting.
}
\details{
bunchit implements the bunching estimator in both kink and notch settings. It bins the running variable, fits a counterfactual density, and estimates the bunching mass (normalized and not), the elasticity and the location of the marginal buncher. In the case of notches, it also finds the dominated region and estimates the fraction of observations located in it.
}
\examples{
# Load the example data before running the examples.
data(bunching_data)

# Example 1: Kink with integration constraint correction
kink1 <- bunchit(z_vector = bunching_data$kink, zstar = 10000,
binwidth = 50, bins_l = 20, bins_r = 20, poly = 4,
bins_excl_l = 0, bins_excl_r = 0, t0 = 0, t1 = .2, p_b = TRUE)
kink1$plot

# Example 2: Kink with further bunching at other level in bandwidth
kink2_vector <- c(bunching_data$kink_vector, rep(10200,540))
kink2 <- bunchit(kink2_vector, zstar = 10000, binwidth = 50,
bins_l = 40, bins_r = 40, poly = 6,
bins_excl_l = 0, bins_excl_r = 0, t0=0, t1=.2, correct = FALSE, p_b=TRUE,
extra_fe = 10200)
kink2$plot

# Example 3: Kink with round number bunching
rn1 <- 500;  rn2 <- 250
bpoint <- 10000
kink3_vector <- c(bunching_data$kink_vector,
                  rep(bpoint + rn1, 270),
                  rep(bpoint + 2*rn1,230),
                  rep(bpoint - rn1,260),
                  rep(bpoint - 2*rn1,275),
                  rep(bpoint + rn2, 130),
                  rep(bpoint + 3*rn2,140),
                  rep(bpoint - rn2,120),
                  rep(bpoint - 3*rn2,135))

kink3 <- bunchit(z_vector = kink3_vector, zstar = bpoint, binwidth = 50,
bins_l = 20, bins_r = 20, poly = 6,
bins_excl_l = 0, bins_excl_r = 0, t0=0, t1=.2, correct = FALSE, p_b=TRUE, p_e = TRUE,
p_freq_msize = 1.5, p_b_ypos = 880, rn = c(250,500))
kink3$plot

# Example 4: Notch
notch <- bunchit(z_vector = bunching_data$notch_vector, zstar = 10000, binwidth = 50,
bins_l = 40, bins_r = 40, poly = 5, bins_excl_l = 0,
t0=0.18, t1=.25, correct = FALSE, notch = TRUE)
notch$plot
}
\seealso{
\code{\link{plot_hist}}
}
