% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bunchit.R
\name{bunchit}
\alias{bunchit}
\title{Run the bunching estimator}
\usage{
bunchit(z_vector, binv = "median", zstar, binwidth, bins_l, bins_r, poly,
  bins_excl_l, bins_excl_r, extra_fe = NA, rn = NA, n_boot = 50,
  correct = TRUE, iter_max = 200, t0, t1, notch = FALSE,
  force_notch = FALSE, p_title = "", p_xtitle = "z_name",
  p_ytitle = "Count", p_maxy = NA, p_axis_title_size = 7,
  p_axis_val_size = 7, p_theme = "bw_light", p_freq_color = "black",
  p_cf_color = "maroon", p_zstar_color = "red", p_freq_size = 0.5,
  p_cf_size = 0.5, p_freq_msize = 1, p_zstar_size = 0.5,
  p_b = TRUE, p_b_xpos = NA, p_b_ypos = NA, p_b_size = 3,
  p_domregion_color = "blue", seed = NA,
  p_domregion_ltype = "longdash")
}
\arguments{
\item{z_vector}{a numeric vector of (unbinned) data to be analysed.}

\item{binv}{a string setting location of zstar within its bin ("min", "max" or "median" value). Default is median.}

\item{zstar}{a numeric value for the the bunching point.}

\item{binwidth}{a numeric value for the width of each bin.}

\item{bins_l}{number of bins to left of zstar to use in analysis.}

\item{bins_r}{number of bins to right of zstar to use in analysis.}

\item{poly}{a numeric value for the order of polynomial for counterfactual fit.}

\item{bins_excl_l}{number of bins to left of zstar to include in bunching region.}

\item{bins_excl_r}{number of bins to right of zstar to include in bunching region.}

\item{extra_fe}{a numeric vector of bin values to control for using fixed effects.}

\item{rn}{a numeric vector of round numbers (up to 2) to control for.}

\item{n_boot}{number of bootstrapped iterations.}

\item{correct}{if TRUE, implements correction for integration constraint.}

\item{iter_max}{maximum iterations for integration constraint correction.}

\item{t0}{numeric value between 0 and 1 setting the marginal/average tax rate below zstar, depending on kink/notch choice. see notch parameter.}

\item{t1}{numeric value between 0 and 1 setting the marginal/average tax rate above zstar, depending on kink/notch choice. see notch parameter.}

\item{notch}{if TRUE, zstar treated as notch. Default is kink.}

\item{force_notch}{if TRUE, user choice of zu (upper limit of bunching region) is enforced. Default is FALSE (zu set by setting bunching equal to missing mass).}

\item{p_title}{plot title.}

\item{p_xtitle}{plot x_axis label.}

\item{p_ytitle}{plot y_axis label.}

\item{p_maxy}{plot's maximum y_axis value.}

\item{p_axis_title_size}{size of plot's axes' title labels.}

\item{p_axis_val_size}{size of plot's axes' numeric labels.}

\item{p_theme}{plot theme (in ggplot2 format).}

\item{p_freq_color}{plot's frequency line color}

\item{p_cf_color}{plot's counterfactual line color}

\item{p_zstar_color}{plot's bunching region marker lines color}

\item{p_freq_size}{plot's frequency line thickness}

\item{p_cf_size}{plot's counterfactual line thickness}

\item{p_freq_msize}{plot's frequency line marker size}

\item{p_zstar_size}{plot's bunching region marker lines thickness}

\item{p_b}{if TRUE, plot also includes bunching estimate. Default is TRUE.}

\item{p_b_xpos}{plot's xaxis coordinate of bunching estimate}

\item{p_b_ypos}{plot's yaxis coordinate of bunching estimate}

\item{p_b_size}{size of plot's printed bunching estimate}

\item{p_domregion_color}{plot's dominated region marker line color (notch case).}

\item{seed}{a numeric value for bootstrap seed (random re-sampling of residuals).}

\item{p_domregion_ltype}{a string for the vertical line type marking the dominated region (zD) in the plot (notch case only).}
}
\value{
A list with the following:
\describe{
  \item{data}{The binned data used for estimation.}
  \item{cf}{The estimated counterfactuals.}
  \item{B}{The estimated excess mass (not normalized).}
  \item{B_vector}{The vector of bootstrapped B's.}
  \item{B_sd}{The standard deviation of B_vector.}
  \item{b}{The estimated excess mass (normalized).}
  \item{b_vector}{The vector of bootstrapped b's.}
  \item{b_sd}{The standard deviation of b_vector.}
  \item{e}{The estimated elasticity.}
  \item{e_vector}{The vector of bootstrapped elasticities (e).}
  \item{e_sd}{The standard deviation of e_vector.}
  \item{alpha}{The estimated fraction of bunchers in dominated region (notch case).}
  \item{alpha_vector}{The vector of bootstrapped alphas.}
  \item{alpha_sd}{The standard deviation of alpha_vector.}
  \item{model_fit}{The model fit on the actual (i.e. not bootstrapped) data.}
  \item{plot}{The bunching plot.}
  \item{zD}{The value demarcating the dominated region (notch case).}
  \item{zD_bin}{The bin above zstar demarcating the dominated region (notch case).}
  \item{marginal_buncher}{The location (z value) of the marginal buncher.}
  \item{marginal_buncher_vector}{The vector of bootstrapped marginal_buncher values.}
  \item{marginal_buncher_sd}{The standard deviation of marginal_buncher_vector.}
}
}
\description{
Run the bunching estimator
}
