% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_bootstrap.R
\name{do_bootstrap}
\alias{do_bootstrap}
\title{Estimate bootstrapped bunching standard error}
\usage{
do_bootstrap(firstpass_prep, residuals, boot_iterations, correction,
  correction_iterations, notch, zD_bin, seed)
}
\arguments{
\item{firstpass_prep}{(binned) data that includes all variables necessary for fitting the model.}

\item{residuals}{residuals from (first pass) fitted bunchingm model}

\item{boot_iterations}{number of bootstrapped samples}

\item{correction}{whether to implement correction for integration constraint.}

\item{correction_iterations}{maximum iterations for integration constraint correction.}

\item{notch}{if TRUE, zstar treated as notch. Default is kink.}

\item{seed}{a numeric value for bootstrap seed (random re-sampling of residuals).}
}
\value{
A list with components...
}
\description{
Estimate bootstrapped bunching standard error
}
\seealso{
\code{\link{bunchit}}, \code{\link{prepare_data}}
}
