% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_bootstrap.R
\name{do_bootstrap}
\alias{do_bootstrap}
\title{Estimate bootstrapped bunching standard error}
\usage{
do_bootstrap(firstpass_prep, residuals, boot_iterations, correction,
  correction_iterations, T0)
}
\arguments{
\item{firstpass_prep}{(binned) data that includes all variables necessary for fitting the model.}

\item{residuals}{residuals from (first pass) fitted bunchingm model}

\item{boot_iterations}{number of bootstrapped samples}

\item{correction}{whether to implement correction for integration constraint.}

\item{correction_iterations}{maximum iterations for integration constraint correction.}

\item{T0}{size of notch (for kink, T0 = 0).}
}
\value{
b_vector, b_sd, B_vector, B_sd
}
\description{
Estimate bootstrapped bunching standard error
}
\seealso{
\code{\link{bunchit}}, \code{\link{prepare_data}}
}
