% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_correction.R
\name{do_correction}
\alias{do_correction}
\title{Implements correction for integration constraint.}
\usage{
do_correction(thedata, firstpass_results, max_iterations, notch)
}
\arguments{
\item{thedata}{(binned) data that includes all variables necessary for fitting the model.}

\item{firstpass_results}{initial bunching estimates without correction.}

\item{max_iterations}{maximum number of iterations for counterfactual shifting.}

\item{notch}{whether it is a notch or kink. Default is kink.}
}
\value{
The data used for estimation, and the "corrected" estimates of both the pure excess mass (B_corrected) and the normalised excess mass (b_corrected).
}
\description{
Implements correction for integration constraint.
}
\seealso{
\code{\link{bunchit}} \code{\link{fit_bunching}}
}
