% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domregion.R
\name{domregion}
\alias{domregion}
\title{Dominated Region}
\usage{
domregion(zstar, t0, t1, binwidth)
}
\arguments{
\item{zstar}{a numeric value for the the bunching point.}

\item{t0}{numeric value between 0 and 1 setting the marginal/average tax rate below zstar, depending on kink/notch choice. see notch parameter.}

\item{t1}{numeric value between 0 and 1 setting the marginal/average tax rate above zstar, depending on kink/notch choice. see notch parameter.}

\item{binwidth}{a numeric value for the width of each bin.}
}
\value{
\code{domregion} returns a list with the following objects related to the dominated region (in notch settings only):
\item{zD}{The level of z that demarcates the upper bound of the dominated region.}
\item{zD_bin}{The value of the bin which zD falls in.}
}
\description{
Estimate z (the value of the running variable) that demarcates the dominated region (in notch settings only).
}
\seealso{
\code{\link{bunchit}}
}
