% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domregion.R
\name{domregion}
\alias{domregion}
\title{Estimate z that demarcates dominated region (in notch settings).}
\usage{
domregion(zstar, t0, t1, binwidth)
}
\arguments{
\item{zstar}{the bunching point.}

\item{t0}{marginal/average tax rate below zstar. see notch option.}

\item{t1}{marginal/average tax rate above zstar. see notch option.}

\item{binwidth}{width of each bin.}
}
\value{
The bin above zstar where individual is indifferent between there or at zstar. Used for notch
}
\description{
Estimate z that demarcates dominated region (in notch settings).
}
