% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domregion.R
\name{domregion}
\alias{domregion}
\title{Estimate z (the value of the running variable) that demarcates the dominated region (in notch settings only).}
\usage{
domregion(zstar, t0, t1, binwidth)
}
\arguments{
\item{zstar}{a numeric value for the the bunching point.}

\item{t0}{numeric value between 0 and 1 setting the marginal/average tax rate below zstar, depending on kink/notch choice. see notch parameter.}

\item{t1}{numeric value between 0 and 1 setting the marginal/average tax rate above zstar, depending on kink/notch choice. see notch parameter.}

\item{binwidth}{a numeric value for the width of each bin.}
}
\value{
The level of z, and the bin it is in (above zstar), that demarcates the dominated region (in notch settings only).
}
\description{
Estimate z (the value of the running variable) that demarcates the dominated region (in notch settings only).
}
