% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticity.R
\name{elasticity}
\alias{elasticity}
\title{Estimate elasticity from single bunching observation.}
\usage{
elasticity(beta, binwidth, zstar, t0, t1, notch)
}
\arguments{
\item{beta}{normalised excess mass.}

\item{binwidth}{width of each bin.}

\item{zstar}{the bunching point.}

\item{t0}{marginal/average tax rate below zstar. see notch option.}

\item{t1}{marginal/average tax rate above zstar. see notch option.}

\item{notch}{whether it is a notch or kink. Default is kink.}
}
\value{
Elasticity estimate.
}
\description{
Estimate elasticity from single bunching observation.
}
\seealso{
\code{\link{bunchit}}
}
