% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticity.R
\name{elasticity}
\alias{elasticity}
\title{Estimate elasticity from single normalised bunching observation.}
\usage{
elasticity(beta, binwidth, zstar, t0, t1, notch)
}
\arguments{
\item{beta}{normalised excess mass.}

\item{binwidth}{a numeric value for the width of each bin.}

\item{zstar}{a numeric value for the the bunching point.}

\item{t0}{numeric value between 0 and 1 setting the marginal/average tax rate below zstar, depending on kink/notch choice. see notch parameter.}

\item{t1}{numeric value between 0 and 1 setting the marginal/average tax rate above zstar, depending on kink/notch choice. see notch parameter.}

\item{notch}{if TRUE, zstar treated as notch. Default is kink.}
}
\value{
Returns the elasticity. In the kink case it uses the standard log-form. In the notch case, it is the reduced-form approximation in equation (5) of Kleven's 2018 note "Calculating Reduced-Form Elasticities Using Notches".
}
\description{
Estimate elasticity from single normalised bunching observation.
}
\seealso{
\code{\link{bunchit}}
}
