% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticity_sd.R
\name{elasticity_sd}
\alias{elasticity_sd}
\title{Estimate s.d. of elasticity.}
\usage{
elasticity_sd(b_vector, binwidth, zstar, t0, t1)
}
\arguments{
\item{b_vector}{vector of normalised excess mass estimates.}

\item{binwidth}{width of each bin.}

\item{zstar}{the bunching point.}

\item{t0}{Marginal tax rate below zstar.}

\item{t1}{Marginal tax rate above zstar.}
}
\value{
Standard deviation of elasticity estimates.
}
\description{
Estimate s.d. of elasticity.
}
\seealso{
\code{\link{elasticity}} \code{\link{bunchit}}
}
