% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticity_sd.R
\name{elasticity_sd}
\alias{elasticity_sd}
\title{Estimate s.d. of elasticity.}
\usage{
elasticity_sd(b_vector, binwidth, zstar, t0, t1, notch)
}
\arguments{
\item{b_vector}{vector of normalised excess mass estimates.}

\item{binwidth}{width of each bin.}

\item{zstar}{the bunching point.}

\item{t0}{marginal/average tax rate below zstar. see notch option.}

\item{t1}{marginal/average tax rate above zstar. see notch option.}

\item{notch}{whether it is a notch or kink. Default is kink.}
}
\value{
Standard deviation of elasticity estimates.
}
\description{
Estimate s.d. of elasticity.
}
\seealso{
\code{\link{elasticity}} \code{\link{bunchit}}
}
