% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_bunching.R
\name{fit_bunching}
\alias{fit_bunching}
\title{Fit  bunching model and estimate excess mass.}
\usage{
fit_bunching(thedata, themodelformula, notch)
}
\arguments{
\item{thedata}{(binned) data that includes all variables necessary for fitting the model.}

\item{themodelformula}{formula to fit.}

\item{notch}{whether it is a notch or kink. Default is kink.}
}
\value{
coefficients, residuals, cf_density, bunchers_excess, cf_bunchers, b_estimate, bins_bunchers, model_formula.
}
\description{
Fit  bunching model and estimate excess mass.
}
\seealso{
\code{\link{bunchit}}, \code{\link{prepare_data}}
}
