% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_buncher.R
\name{marginal_buncher}
\alias{marginal_buncher}
\title{Marginal Buncher}
\usage{
marginal_buncher(beta, binwidth, zstar)
}
\arguments{
\item{beta}{normalised excess mass.}

\item{binwidth}{a numeric value for the width of each bin.}

\item{zstar}{a numeric value for the the bunching point.}
}
\value{
\code{marginal_buncher} returns the location of the marginal buncher, i.e. zstar + Dzstar.
}
\description{
Calculate location (value of running variable) of marginal buncher.
}
\seealso{
\code{\link{bunchit}}
}
