% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bunching.R
\name{plot_bunching}
\alias{plot_bunching}
\title{Create bunching plot.}
\usage{
plot_bunching(binned_data, cf, zstar, binwidth, bins_excl_l, bins_excl_r,
  p_title, p_xtitle, p_ytitle, p_maxy, p_axis_title_size, p_axis_val_size,
  p_theme, p_freq_color, p_cf_color, p_zstar_color, p_freq_size, p_cf_size,
  p_freq_msize, p_zstar_size, p_b, b, b_sd, p_b_xpos, p_b_ypos, p_b_size,
  t0 = NA, t1 = NA, notch = F, domregion_color = NA,
  domregion_ltype = NA)
}
\arguments{
\item{binned_data}{binned data with frequency and counterfactual estimated}

\item{zstar}{a numeric value for the the bunching point.}

\item{binwidth}{a numeric value for the width of each bin.}

\item{bins_excl_l}{number of bins to left of zstar to include in bunching region.}

\item{bins_excl_r}{number of bins to right of zstar to include in bunching region.}

\item{p_title}{plot title.}

\item{p_xtitle}{plot x_axis label.}

\item{p_ytitle}{plot y_axis label.}

\item{p_maxy}{plot's maximum y_axis value.}

\item{p_axis_title_size}{size of plot's axes' title labels.}

\item{p_axis_val_size}{size of plot's axes' numeric labels.}

\item{p_theme}{plot theme (in ggplot2 format).}

\item{p_freq_color}{plot's frequency line color}

\item{p_cf_color}{plot's counterfactual line color}

\item{p_zstar_color}{plot's bunching region marker lines color}

\item{p_freq_size}{plot's frequency line thickness}

\item{p_cf_size}{plot's counterfactual line thickness}

\item{p_freq_msize}{plot's frequency line marker size}

\item{p_zstar_size}{plot's bunching region marker lines thickness}

\item{p_b}{if TRUE, plot also includes bunching estimate. Default is TRUE.}

\item{p_b_xpos}{plot's xaxis coordinate of bunching estimate}

\item{p_b_ypos}{plot's yaxis coordinate of bunching estimate}

\item{p_b_size}{size of plot's printed bunching estimate}

\item{t0}{numeric value between 0 and 1 setting the marginal/average tax rate below zstar, depending on kink/notch choice. see notch parameter.}

\item{t1}{numeric value between 0 and 1 setting the marginal/average tax rate above zstar, depending on kink/notch choice. see notch parameter.}

\item{notch}{if TRUE, zstar treated as notch. Default is kink.}
}
\value{
A plot with the frequency, counterfactual and bunching region demarcated. Can also include the bunching estimate in the plot if specified.
}
\description{
Create bunching plot.
}
\seealso{
\code{\link{bunchit}}
}
